requirements.txt
torch
torchvision
fastapi
uvicorn[standard]
opencv-python
numpy
pandas
Pillow
python-multipart